refactor: move chunks function and key type into separate files in better locations
