+++ src/api/types.ts
@@ -1,0 +1,8 @@
+export type Key = {
+  itemId: string;
+  publicKey: string;
+  role: string;
+  userId: string;
+  itemType: string;
+  key: string;
+};
